KWalls Simple Wallpaper App
Scroll through a list of wallpapers and select a wallpaper that you
want. Once selected the wallpaper will be you new background. Based on
the wallpaper you select you can choose to download or share the image.
The image that is downloaded will be located in your photos. There is
also a link to my GitHub account also on the app once the user selects a
button.